Fix review comments
- Rename attr_coverage functions to be more descriptive
- Move tests to the file that they test
- Fix build errors caused by changes to the Event enum
